+++ src/app/functions/forms/index.ts
@@ -1,0 +1,9 @@
+/**
+ * フォーム検証
+ */
+
+export * from './inquiry/login';
+export * from './purchase/input';
+export * from './purchase/mvtk/mvtk-input';
+export * from './purchase/seat';
+export * from './purchase/ticket';