+++ src/app/functions/forms/purchase/mvtk/mvtk-input.ts
@@ -1,0 +1,15 @@
+import * as express from 'express';
+
+/**
+ * ムビチケ認証
+ */
+export function purchaseMvtkInputForm (req: express.Request): void {
+    req.checkBody(
+        'mvtk',
+        `${req.__('common.mvtk')}${req.__('common.validation.required')}`
+    ).notEmpty();
+    req.checkBody(
+        'mvtk',
+        `${req.__('common.mvtk')}${req.__('common.validation.is_json')}`
+    ).isJSON();
+}