+++ src/app/models/index.ts
@@ -1,0 +1,8 @@
+/**
+ * モデル
+ */
+
+export * from './auth';
+export * from './inquiry';
+export * from './purchase';
+export * from './error';