+++ CHANGELOG.md
@@ -12,6 +12,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## 2.5.0 - 2019-05-24
+### Changed
+- [@motionpicture/sskts-api-javascript-client@8.0.1](https://github.com/motionpicture/sskts-api-javascript-client)に対応。
+- [@motionpicture/sskts-api-nodejs-client@8.0.1](https://github.com/motionpicture/sskts-api-nodejs-client)に対応。
+
 ## 2.4.2 - 2019-02-08
 ### Fixed
  - 印刷不具合修正（APIアップデートによる互換性調整）